# contracts/http/openapi.yaml
...
openapi: 3.1.0
info:
  title: Research Radar — HTTP wire contract
  version: 0.3.0
  summary: 'Hợp đồng HTTP cho toàn bộ operation `transport: http` của contracts/ports.yaml.'
  description: |-
    Đây là hợp đồng dây (wire) của Research Radar, sinh từ `contracts/ports.yaml` và giữ nguyên `operation_id`
    làm `operationId`. Mỗi operation `transport: http` trong ports.yaml có ĐÚNG MỘT cặp path+method ở đây.

    Những gì file này KHÔNG định nghĩa, và nơi chúng thuộc về:
      - Ngữ nghĩa transition trạng thái  → contracts/state/*.yaml (PC03)
      - Danh mục mã lỗi và envelope      → contracts/errors.yaml (PC03); `ErrorEnvelope` dưới đây là bản sao ĐÚNG
                                           hình dạng của `error_envelope` trong file đó
      - Ngân sách retry/timeout/lease    → contracts/retry-policy.yaml (PC03)
      - Ranh giới module và cạnh bị cấm  → contracts/modules.yaml (PC01)
      - Nội dung/coverage báo cáo        → contracts/reporting/* (PC04)
      - Task và schema AI                → contracts/ai/* (PC06)
      - Ngữ nghĩa lệnh Telegram          → contracts/telegram/commands.yaml (PC07)
      - Vòng đời secret, CSRF, xoay token→ contracts/ops/secrets.md (PC08)

    Wire rules bắt buộc (SRC-PLAN §5.1), áp cho mọi operation:
      - ID truyền dưới dạng string; timestamp UTC RFC 3339 với độ chính xác mili giây (AMD-B08).
      - Mutation từ worker mang `request_id`, `schema_version`, `idempotency_key`, `payload_hash`; mutation đang
...

...
  x-contract:
...
    - REQ-D04
...
servers:
...
security:
- ownerSessionCookie: []
  ownerCsrfToken: []
components:
  securitySchemes:
    ownerSessionCookie:
      type: apiKey
      in: cookie
      name: rr_session
      description: 'Cookie phiên HttpOnly + Secure + SameSite=Lax. PROVISIONAL, phối hợp với PC08 (xem info.x-contract.security_scheme_decision).
        Với MUTATION, scheme này xuất hiện CÙNG `ownerCsrfToken` trong MỘT security requirement — tức quan hệ AND, cookie
        một mình KHÔNG đủ. Với read-only, cookie đứng một mình: CSRF bảo vệ thao tác đổi trạng thái, không phải phép đọc.'
    ownerCsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 'CSRF double-submit: giá trị phải khớp cookie `rr_csrf` (không-HttpOnly). Bắt buộc cho MỌI mutation của
        owner session. Thiếu hoặc lệch ⇒ 403 `CSRF_REJECTED` (KHÔNG phải `FORBIDDEN_EDGE`: cạnh vẫn hợp lệ, chỉ lời gọi
        không chứng minh được nó do trang của ứng dụng phát ra — ruling CR-PC08-03/CR-PC08-04). Phiên KHÔNG bị hủy khi gặp
        mã này.'
    collectorToken:
      type: http
      scheme: bearer
      description: Token của collector, lưu trong file cấu hình QUYỀN HẠN CHẾ trên máy cá nhân, KHÔNG BAO GIỜ trong repo
        (SRC-SPEC §11.2). Chỉ mở được các operation `worker.*` và `ingest.*`; gọi `save.create` bằng token này ⇒ 401 (denied
        case NC-01).
    analysisWorkerToken:
      type: http
      scheme: bearer
      description: Token của analysis worker. Chỉ mở `analysis.*` và `secret.issue_task_credential`. KHÔNG mở đường tới
        tag, report hay Saved (denied case NC-02).
    telegramIngressSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Bí mật webhook do app đặt khi đăng ký webhook với Telegram. Sai hoặc thiếu ⇒ 401 và KHÔNG xử lý update.
        Biết chat ID KHÔNG phải là quyền (SRC-SPEC §11.3).
    backupOperatorToken:
      type: http
      scheme: bearer
      description: 'Danh tính operator backup. Tách khỏi phiên owner vì restore là thao tác có khóa side effect (AMD-B11).
        Sửa ở FIX1 (CR-PC08-02): mọi route `backup.*` nhận DUY NHẤT scheme này — trước đây phiên trình duyệt cũng qua được,
        nghĩa là một tab đang mở có thể kích hoạt restore.'
  parameters:
    SchemaVersionHeader:
      name: X-Schema-Version
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/SemVer'
      description: Phiên bản wire schema của request. KHÔNG dùng 'latest' (SRC-PLAN §5). Lệch major ⇒ 422 `VALIDATION_ERROR`
        với `details_safe.violation_kind = schema_version_unsupported`.
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Ulid'
      description: ID của LẦN GỌI này. Một retry sinh `X-Request-Id` MỚI nhưng GIỮ NGUYÊN `Idempotency-Key` — đó là cách
        server phân biệt replay với một yêu cầu mới.
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/IdempotencyKey'
      description: Khóa chống trùng do client cấp; phạm vi khóa nằm ở `x-idempotency.key_scope` của từng operation. Cùng
        key + cùng payload_hash ⇒ trả kết quả đã commit. Cùng key + payload khác ⇒ 409 `IDEMPOTENCY_CONFLICT`, KHÔNG ghi
        đè (SRC-PLAN §5.1).
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      description: |-
        Envelope lỗi chung (SRC-PLAN §5.1). Hình dạng này là BẢN SAO của `error_envelope` trong contracts/errors.yaml;
        nếu hai bên lệch nhau thì contracts/errors.yaml THẮNG và file này phải sửa.

        `correlation_id` LUÔN có mặt, kể cả khi `storage.health != healthy`: nó sinh trong bộ nhớ tiến trình, không cần
        một hàng DB. Không mã lỗi nào đòi hỏi ghi được vào DB mới coi là đã báo (contracts/errors.yaml §EPR-01).

        Envelope KHÔNG BAO GIỜ chứa: transcript của model, prompt gốc, nội dung post nguyên văn, API key, bot token,
        cookie, header Authorization, chat ID, đường dẫn file tuyệt đối, stack trace, câu SQL, tên bảng/cột nội bộ.
      required:
      - code
      - scope
      - retry_class
      - message_safe
      - correlation_id
      properties:
        code:
          type: string
          enum:
          - X_CHALLENGE_REQUIRED
          - X_ACCESS_BLOCKED
          - WORKER_LEASE_EXPIRED
          - STALE_LEASE
          - INGEST_ACK_LOST
          - IDEMPOTENCY_CONFLICT
          - SOURCE_METADATA_UNAVAILABLE
          - IDENTITY_CONFLICT
          - AI_OUTPUT_INVALID
          - AI_PROVIDER_UNAVAILABLE
          - AI_ATTEMPT_UNCERTAIN
          - EMBEDDING_GENERATION_MISMATCH
          - TAG_VERSION_STALE
          - CONFLICT
          - TELEGRAM_SEND_UNCERTAIN
          - TELEGRAM_PERMANENT_FAILURE
          - UNAUTHORIZED_COMMAND
          - STORAGE_WRITE_FAILED
          - RESTORE_UNVERIFIED
          - SOURCE_LAYOUT_CHANGED
          - CSRF_REJECTED
          - UNAUTHORIZED
          - CAPABILITY_DENIED
          - FORBIDDEN_EDGE
          - VALIDATION_ERROR
          - NOT_FOUND
          - RATE_LIMITED
          - INTERNAL
          description: Phải là một mã đã khai trong contracts/errors.yaml. Không có mã ngoài danh mục.
        scope:
          type: string
          enum:
          - run
          - item
          - delivery
          - storage
          - request
          description: Bằng đúng `scope` của mã trong contracts/errors.yaml.
        retry_class:
          type: string
          enum:
          - none
          - retryable_with_budget
          - needs_user
          - operator_decision
          - unknown_outcome
          description: Bằng đúng `retry_class` của mã. `unknown_outcome` nghĩa là KHÔNG BIẾT side effect đã xảy ra hay chưa
            — client KHÔNG được coi là rollback.
        message_safe:
          type: string
          maxLength: 300
          description: Tiếng Việt, render từ `message_safe_template` của mã. KHÔNG chèn giá trị ngoài `details_safe_keys`.
        correlation_id:
          $ref: '#/components/schemas/Ulid'
        details_safe:
          type:
          - object
          - 'null'
          description: Chỉ chứa các khóa liệt kê ở `details_safe_keys` của mã trong contracts/errors.yaml. Khóa lạ ⇒ producer
            tự từ chối, không gửi ra.
        retry_after_ms:
          type:
          - integer
          - 'null'
          minimum: 0
          description: Chỉ có mặt khi `retry_class = retryable_with_budget` hoặc mã `RATE_LIMITED`. null nghĩa là KHÔNG
            có gợi ý, KHÔNG phải 'retry ngay'.
    Ulid:
      type: string
      pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
      description: ID truyền dưới dạng string (SRC-PLAN §5.1).
    SemVer:
      type: string
      pattern: ^[0-9]+\.[0-9]+\.[0-9]+$
...
    IdempotencyKey:
      type: string
      pattern: ^[A-Za-z0-9_.:-]{16,128}$
...
    GenericObject:
      type: object
      description: |-
        Chỗ giữ có chủ đích: hình dạng chi tiết thuộc một gói khác và CHƯA tồn tại tại thời điểm viết file này.
        Xem `$ref` trong `x-response-schema-planned` của ports.yaml cho từng operation. Đây KHÔNG phải giấy phép chấp
        nhận trường tùy ý: gói sở hữu phải thay bằng schema đóng với `additionalProperties: false` trước G3.
...
paths:
...
  /v1/analysis/tasks/claim:
    post:
      operationId: analysis.claim_task
      summary: Analysis worker nhận task từ server
      description: 'worker_instance_id, capability provider hiện có, số task muốn nhận.


        analysis `pending` → `running`; cấp lease/epoch.


        Task CHỈ được tạo cho dữ liệu ĐÃ COMMIT ở server (B12/AMD-B12); worker không nhận dữ liệu trực tiếp từ collector.
        Schema chi tiết thuộc PC06 (`contracts/ai/tasks.yaml`).'
      tags:
      - analysis
      security:
      - analysisWorkerToken: []
      x-auth-scope: analysis_worker_token
      x-owner-module: MOD-analysis-service
      x-caller-modules:
      - MOD-analysis-worker
      x-mutation: true
      x-scenario-refs:
      - SC28
      x-requirement-refs:
      - REQ-D42
      - REQ-P0-11
      x-decision-refs:
      - B07
      - B12
      - B13
      - ADR-0001
      x-invariant-refs:
      - I01
      - I04
      x-idempotency:
        required: true
        key_scope: claim_request_id
        rule_vi: Cùng claim_request_id trả cùng task; không cấp hai lease cho một task.
        header: Idempotency-Key
      parameters:
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: task_id, task type, analysis_key, lease_id, lease_epoch, tham chiếu input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericObject'
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 'Lỗi: `NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'Lỗi: `IDEMPOTENCY_CONFLICT`, `STALE_LEASE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'Lỗi: `VALIDATION_ERROR`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 'Lỗi: `RATE_LIMITED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây do nguồn/API chỉ định. Giá trị này THẮNG mọi backoff cấu hình.
              schema:
                type: integer
                minimum: 0
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: 'Lỗi: `STORAGE_WRITE_FAILED`. `storage.health = write_blocked` hoặc `maintenance`: mutation KHÔNG
            được ACK và KHÔNG có gì được ghi (SRC-PLAN §8.4, I02).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây. Gợi ý theo `storage_recovery_probe` (30 s).
              schema:
                type: integer
                minimum: 0
...
  /v1/secrets/task-credentials:
    post:
      operationId: secret.issue_task_credential
      summary: Cấp credential ngắn hạn theo từng task cho analysis worker (B13)
      description: 'task_id + lease_id/epoch.


        Ghi audit cấp phát đã che secrets; hạn dùng do PC08 chốt.


        Credential có PHẠM VI ĐÚNG MỘT TASK và thời hạn ngắn (B13/ADR-0010). KHÔNG bao giờ cấp toàn bộ secrets. Giá trị
        credential KHÔNG được ghi log. Hình dạng chi tiết thuộc PC08.'
      tags:
      - secrets
      security:
      - analysisWorkerToken: []
      x-auth-scope: analysis_worker_token
      x-owner-module: MOD-secret-service
      x-caller-modules:
      - MOD-analysis-worker
      x-mutation: true
      x-scenario-refs:
      - SC16
      - SC17
      x-requirement-refs:
      - REQ-S11.2-01
      - REQ-D51
      x-decision-refs:
      - B13
      x-invariant-refs:
      - I01
      - I11
      x-idempotency:
        required: true
        key_scope: task_id + attempt_id
        rule_vi: Cùng attempt trả cùng credential còn hạn; không cấp credential cho task worker không giữ lease.
        header: Idempotency-Key
      parameters:
      - $ref: '#/components/parameters/SchemaVersionHeader'
      - $ref: '#/components/parameters/RequestIdHeader'
      - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenericObject'
      responses:
        '200':
          description: Credential **chỉ của provider gắn với task đó**, có hạn ngắn. Không trả toàn bộ key store, không
            trả key của provider khác.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/GenericObject'
                x-schema-source: contracts/ops/secrets.md
                description: Hình dạng do hợp đồng `contracts/ops/secrets.md` sở hữu (không phải JSON Schema nên không $ref
                  được). Gói sở hữu phải thay bằng schema đóng trước G3.
          headers:
            X-Schema-Version:
              description: Phiên bản schema của response.
              schema:
                $ref: '#/components/schemas/SemVer'
        '401':
          description: 'Lỗi: `UNAUTHORIZED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 'Lỗi: `FORBIDDEN_EDGE`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: 'Lỗi: `NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 'Lỗi: `STALE_LEASE`, `IDEMPOTENCY_CONFLICT`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: 'Lỗi: `VALIDATION_ERROR`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: 'Lỗi: `RATE_LIMITED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          headers:
            Retry-After:
              description: Giây do nguồn/API chỉ định. Giá trị này THẮNG mọi backoff cấu hình.
              schema:
                type: integer
                minimum: 0
        '500':
          description: 'Lỗi: `INTERNAL`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...